Carrot_12_20

+++ workshops/models.py
@@ -27,7 +27,7 @@ class Workshop(models.Model):
     content = models.TextField(max_length=500)
     workshop_image = models.ImageField(upload_to=rename_imagefile_to_uuid)
     max_guest = models.IntegerField()
-    amount = models.DecimalField(decimal_places=0, max_digits=10000000000000000)
+    amount = models.PositiveIntegerField()
     created_at = models.DateTimeField(auto_now_add=True)
     participant = models.ManyToManyField(User, related_name='workshop_participant', blank=True, through='WorkshopApply', through_fields=('workshop', 'guest'),) 
     likes = models.ManyToManyField(User, related_name='workshop_likes', blank=True)
